[AN] refactor: 검색 컴포넌트 파일 구조 통합 #797
* [AN] refactor: HearitMetaRow.kt 및 HearitTitle.kt 파일을 삭제하고 해당 컴포넌트들을 HearitItem.kt로 이동
* [AN] refactor: HearitItem.kt에 관련 import 및 하위 컴포넌트(HearitTitle, HearitMetaRow) 로직 추가
